feat: ✨ Mod Hooks (#408)
* feat: :construction: first draft of the export plugin

The plan is to automate the creation of hooks for each function in all GDScripts.

* feat: :construction: wip callable_stack

* refactor: :construction: rename vanilla functions

Vanilla functions are renamed, and at the bottom of the script, mod loader hooks are generated. These hooks replace the vanilla functions and handle calls to the `callable_stack`.

* feat: :construction: added return with types to hook

Currently, I'm stuck on this approach because the method list's return property does not contain any information if there is no return type defined for the method.

* refactor: :construction: reworked callable_sack funcs

* refactor: :construction: added return usage check

That seems to work for filtering out methods that have a return.

* refactor: :construction: remove and add back class_name

Without removing the `class_name` before initializing the script on export, I get this parse error: `hides a global script class`. With these code changes, I get `Value of type [script_path] cannot be assigned to a variable of type [class_name]` if there is a self-assignment somewhere in the script.

* refactor: :construction: No new script instance - load() the script 😶‍🌫️

Also added a check for static functions and currently working on recognizing inner classes.

* refactor: :construction: added `get_return_type_string`

check the `type_hint` and add it if it's present.

* refactor: :construction: added `is_top_level_func`

To check if the function has no space before `func` or `static`, used to ignore functions in inner classes. Later, we might want to figure out a way to include these.

* fix: :construction: corrected return type in generated methods

Also removed extra comma in args list

* feat: :construction: only add modloader text to modified scripts

until now all scripts received the "Mod Loader Hooks" header at the bottom
even when no changes where made.

* feat: :construction: added optional @moddable requirement

* feat: :construction: added `is_setter`

used to ignore setter funcs

* feat: :construction: improved runtime performance by pre-hashing dict-lookup

Also refactored some of the function/field names to improved the modding experience

* feat: :construction: fixed issues with inheritance by adding class hash

added getter method, made both static and moved getter/setter check to is_moddable check

* cleaner getters and setters

* feat: :construction: added @not-moddable option to exclude methods

improved performance by exiting early when nothing is hooked

* refactor: :recycle: removed no longer used functions

and a first style and type pass

* refactor: 🚧 general code cleanup

* :construction: refactor: move methods

* :construction: refactor: extract methods

* :construction: feat: mod hook packing prototype

* fix: :construction: fix path to `mod_hook_preprocessor.gd`

* :construction: fix: typos

* feat: :construction: replace super calls

* refactor: :recycle: remove static from regex vars

* feat: :construction: hook creation and loading (#3)

* feat: :construction: hook creation and loading

* feat: :sparkles: hook creation and loading

---------

Co-authored-by: Luca Martinelli <lucxmangajet@gmail.com>
Co-authored-by: Qubus0 <steen.rickmer@gmx.de>