refactor: :recycle: new self setup for Godot 4
Switched from PckTool to PckExplorer. This allows us to extract embedded .pck files from the game .exe and inject the modified project.binary and global_script_class_cache.cfg. This ensures that the mod loader autoloads are at the top of the load order, which is crucial for hooking into or extending game autoloads.