fix: :bug: match any character between func name and `(` (#495)
* fix: :bug: match any character between func name and `(`

* refactor: :recycle: use non-capturing group

* fix: :bug: match white space before `(`

* fix: Ignore comments

* refactor: :fire: removed redundant whitespace token

moved the regex string to the top and added a description

* fix: :bug: match space before `(`

and word boundary for the method name

* refactor: :recycle: removed `(?<!#)`

It only handled one specific case, `#func`. Something like `# ` was still matched. To avoid giving a false impression, I removed it, and this case is now also handled by `is_top_level_func()`.

* fix: :bug: more word boundaries