refactor: :recycle: new self setup for Godot 4 (#507)
* refactor: :recycle: new self setup for Godot 4

Switched from PckTool to PckExplorer. This allows us to extract embedded .pck files from the game .exe and inject the modified project.binary and global_script_class_cache.cfg. This ensures that the mod loader autoloads are at the top of the load order, which is crucial for hooking into or extending game autoloads.

* refactor: :fire: removed output print in split_pck()

* refactor: :recycle: set GodotPCKExplorer path based on OS

* refactor: :truck: move pckexporer outside of win dir

* refactor: :fire: remove GodotPCKExplorer

Include only the correct OS version of GodotPCKExplorer in the zip file provided on the release page.

* Update addons/mod_loader/mod_loader_setup.gd

Co-authored-by: steen <steen.rickmer@gmx.de>

---------

Co-authored-by: steen <steen.rickmer@gmx.de>